requirements.txt
streamlit
sounddevice
scipy
requests
python-dotenv
PyMuPDF  # for extracting text from resume PDF
pygame